move commands alongside terraform commands
