+++ TerraformCLI/src/az-storage-account-create.ts
@@ -1,10 +1,10 @@
-import { IHandleCommand, ICommand } from "../../commands";
-import { AzureCLI } from "../azure-cli";
-import { Step, StepFrom } from "../step";
+import { IHandleCommand, ICommand } from "./commands";
+import { CommandRunner } from "./command-runner";
+import { Step, StepFrom } from "./step";
 import { injectable, inject } from "inversify";
-import { CommandPipeStep } from "../command-pipe-step";
+import { CommandPipeStep } from "./command-pipe-step";
 
-declare module "../step" {
+declare module "./step" {
     interface Step<TResult> {
         azStorageAccountCreate<TPreviousResult>(this: Step<TPreviousResult>, command: AzStorageAccountCreate): StepFrom<TPreviousResult, AzStorageAccountCreateResult>;
     }
@@ -44,10 +44,10 @@ export class AzStorageAccountCreate implements ICommand<AzStorageAccountCreateRe
 
 @injectable()
 export class AzStorageAccountCreateHandler implements IHandleCommand<AzStorageAccountCreate, AzStorageAccountCreateResult>{
-    private readonly cli: AzureCLI;
+    private readonly cli: CommandRunner;
 
     constructor(
-        @inject(AzureCLI) cli: AzureCLI) {
+        @inject(CommandRunner) cli: CommandRunner) {
         this.cli = cli;
     }
