+++ TerraformCLI/src/azure-shell.ts
