+++ TerraformCLI/src/pipeline-command.ts
@@ -1,8 +1,8 @@
 import { Step } from "./step";
-import { IMediator } from "../mediator";
-import { ICommand } from "../commands";
+import { IMediator } from "./mediator";
+import { ICommand } from "./commands";
 
-export class CommandStep<TCommand extends ICommand<TResult>, TResult> extends Step<TResult>
+export class PipelineCommand<TCommand extends ICommand<TResult>, TResult> extends Step<TResult>
 {
     private readonly command: TCommand;
